LayoutTests/fast/frames/sandboxed-iframe-scripting.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    window.testRunner.dumpAsText();

var disallowedExecuted = 0;
var allowedExecuted = 0;

window.onload = function()
{
    shouldBe("disallowedExecuted", "0");
    shouldBe("allowedExecuted", "3");

    isSuccessfullyParsed();
}
</script>
</head>

<body>
 
    <iframe sandbox="allow-same-origin allow-scripts"
            src="javascript: ++window.top.allowedExecuted;">
    </iframe>
    <iframe sandbox="allow-same-origin allow-scripts"
            src="data:text/html,<script> alert('PASS: Executed script in data URL'); </script>">
    </iframe>
    <iframe sandbox="allow-same-origin"
            src="javascript: ++window.top.disallowedExecuted;">
    </iframe>
    <iframe sandbox="allow-same-origin"
            src="data:text/html,<script> alert('FAIL: Executed script without allow-scripts in data URL'); </script>">
    </iframe>

    <iframe id="frame" src="resources/sandboxed-iframe-script-dynamic.html">
    </iframe>

    <script>
    description("Verify that sandboxed frames with sandbox=&quot;allow-scripts&quot; can execute scripts, "
                 + "but other sandboxed frames cannot. Also verify that adding a sandbox attribute at runtime "
                 + "does not effect the current state of the iframe.");
    </script>
</body>
</html>
