LayoutTests/fast/frames/sandboxed-iframe-scripting-expected.txt

ALERT: PASS: Executed script in data URL
Verify that sandboxed frames with sandbox="allow-scripts" can execute scripts, but other sandboxed frames cannot. Also verify that adding a sandbox attribute at runtime does not effect the current state of the iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS disallowedExecuted is 0
PASS allowedExecuted is 3
PASS successfullyParsed is true

TEST COMPLETE
      
